app-mysql/requirements.txt
mysql-connector-python>=8.0.33
streamlit>=1.28.0
pandas>=2.0.0
streamlit-option-menu>=0.3.5
python-dateutil>=2.8.2
plotly>=5.0.0
